Use spaces instead of indentations
Fix syntax errors

Standardize format

Move UPC up one level
