+++ project/settings.py
@@ -124,7 +124,6 @@ INSTALLED_APPS = [
     'providers.edu.opensiuc',
     'providers.edu.pcom',
     'providers.edu.pdxscholar',
-    'providers.edu.pe.upc',
     'providers.edu.purdue',
     'providers.edu.richmond',
     'providers.edu.scholarsarchiveosu',
@@ -213,6 +212,7 @@ INSTALLED_APPS = [
     'providers.org.ucescholarship',
     'providers.org.zenodo',
     'providers.org.seafdec',
+    'providers.pe.upc',
     'providers.pt.rcaap',
     'providers.ru.cyberleninka',
     'providers.tr.edu.hacettepe',
