--- providers/edu/pe/upc/__init__.py
@@ -1,1 +1,0 @@
-default_app_config = 'providers.edu.pe.upc.apps.AppConfig'
