--- providers/edu/pe/upc/apps.py
@@ -1,10 +1,0 @@
-from share.provider import OAIProviderAppConfig
-
-
-class AppConfig(OAIProviderAppConfig):
-	name = 'providers.edu.pe.upc'
-	version = '0.0.1'
-	title = 'Universidad Peruana de Ciencias Aplicadas (UPC)'
-	long_title = 'Universidad Peruana de Ciencias Aplicadas (UPC)'
-	home_page = 'http://repositorioacademico.upc.edu.pe'
-	url = 'http://repositorioacademico.upc.edu.pe/upc/oai/request'
