+++ providers/edu/richmond/__init__.py
@@ -3,27 +3,27 @@ default_app_config = 'providers.edu.richmond.apps.AppConfig'
 '''
 Example Record:
 <record>
-	<header>
-		<identifier>oai:scholarship.richmond.edu:homepage-1004</identifier>
-		<datestamp>2013-07-31T20:37:11Z</datestamp>
-		<setSpec>publication:homepage</setSpec>
-	</header>
-	<metadata>
-		<oai_dc:dc xmlns:oai_dc="http://www.openarchives.org/OAI/2.0/oai_dc/" xmlns:dc="http://purl.org/dc/elements/1.1/" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.openarchives.org/OAI/2.0/oai_dc/ http://www.openarchives.org/OAI/2.0/oai_dc.xsd">
-		<dc:title>UR Scholarship Slideshow</dc:title>
-		<dc:date>2013-07-31T07:00:00Z</dc:date>
-		<dc:type>text</dc:type>
-		<dc:format>image/jpg</dc:format>
-		<dc:identifier>http://scholarship.richmond.edu/homepage/3</dc:identifier>
-		<dc:identifier>
-		http://scholarship.richmond.edu/context/homepage/article/1004/type/native/viewcontent
-		</dc:identifier>
-		<dc:source>Homepage Slideshow</dc:source>
-		<dc:publisher>UR Scholarship Repository</dc:publisher>
-		<dc:description>
-		http://scholarship.richmond.edu/homepage/1004/thumbnail.jpg
-		</dc:description>
-		</oai_dc:dc>
-	</metadata>
+    <header>
+        <identifier>oai:scholarship.richmond.edu:homepage-1004</identifier>
+        <datestamp>2013-07-31T20:37:11Z</datestamp>
+        <setSpec>publication:homepage</setSpec>
+    </header>
+    <metadata>
+        <oai_dc:dc xmlns:oai_dc="http://www.openarchives.org/OAI/2.0/oai_dc/" xmlns:dc="http://purl.org/dc/elements/1.1/" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.openarchives.org/OAI/2.0/oai_dc/ http://www.openarchives.org/OAI/2.0/oai_dc.xsd">
+        <dc:title>UR Scholarship Slideshow</dc:title>
+        <dc:date>2013-07-31T07:00:00Z</dc:date>
+        <dc:type>text</dc:type>
+        <dc:format>image/jpg</dc:format>
+        <dc:identifier>http://scholarship.richmond.edu/homepage/3</dc:identifier>
+        <dc:identifier>
+            http://scholarship.richmond.edu/context/homepage/article/1004/type/native/viewcontent
+        </dc:identifier>
+        <dc:source>Homepage Slideshow</dc:source>
+        <dc:publisher>UR Scholarship Repository</dc:publisher>
+        <dc:description>
+        http://scholarship.richmond.edu/homepage/1004/thumbnail.jpg
+        </dc:description>
+        </oai_dc:dc>
+    </metadata>
 </record>
 '''
