+++ providers/edu/scholarworks_montana/__init__.py
@@ -1,17 +1,28 @@
 default_app_config = 'providers.edu.scholarworks_montana.apps.AppConfig'
 
 """
-<oai_dc:dc xmlns:dc="http://purl.org/dc/elements/1.1/" xmlns:doc="http://www.lyncode.com/xoai" xmlns:oai_dc="http://www.openarchives.org/OAI/2.0/oai_dc/" xsi:schemaLocation="http://www.openarchives.org/OAI/2.0/oai_dc/ http://www.openarchives.org/OAI/2.0/oai_dc.xsd">
-<dc:title>Confluence, Volume 7, 2010-2011</dc:title>
-<dc:creator>Multiple Authors</dc:creator>
-<dc:subject>College of Letters and Science</dc:subject>
-<dc:description>Confluence is the annual magazine of the College of Letters &amp; Science at Montana State University and is published once per year to highlight the accomplishments of the college. The magazine is distributed to alumni, friends and prospective students.</dc:description>
-<dc:date>2012-09-24T21:20:27Z</dc:date>
-<dc:date>2012-09-24T21:20:27Z</dc:date>
-<dc:date>2010</dc:date>
-<dc:type>Other</dc:type>
-<dc:identifier>http://scholarworks.montana.edu/xmlui/handle/1/315</dc:identifier>
-<dc:language>en_US</dc:language>
-<dc:publisher>College of Letters and Science</dc:publisher>
+Example Record
+
+<oai_dc:dc xmlns:dc="http://purl.org/dc/elements/1.1/"
+    xmlns:doc="http://www.lyncode.com/xoai"
+    xmlns:oai_dc="http://www.openarchives.org/OAI/2.0/oai_dc/"
+    xsi:schemaLocation="http://www.openarchives.org/OAI/2.0/oai_dc/
+    http://www.openarchives.org/OAI/2.0/oai_dc.xsd">
+    <dc:title>Confluence, Volume 7, 2010-2011</dc:title>
+    <dc:creator>Multiple Authors</dc:creator>
+    <dc:subject>College of Letters and Science</dc:subject>
+    <dc:description>
+        Confluence is the annual magazine of the College of Letters &amp; Science at
+        Montana State University and is published once per year to highlight the
+        accomplishments of the college. The magazine is distributed to alumni, friends
+        and prospective students.
+    </dc:description>
+    <dc:date>2012-09-24T21:20:27Z</dc:date>
+    <dc:date>2012-09-24T21:20:27Z</dc:date>
+    <dc:date>2010</dc:date>
+    <dc:type>Other</dc:type>
+    <dc:identifier>http://scholarworks.montana.edu/xmlui/handle/1/315</dc:identifier>
+    <dc:language>en_US</dc:language>
+    <dc:publisher>College of Letters and Science</dc:publisher>
 </oai_dc:dc>
 """
