+++ providers/org/ttu/__init__.py
@@ -1,35 +1,43 @@
 default_app_config = 'providers.org.ttu.apps.AppConfig'
+
 """
-<oai_dc:dc xmlns:dc="http://purl.org/dc/elements/1.1/" xmlns:doc="http://www.lyncode.com/xoai" xmlns:oai_dc="http://www.openarchives.org/OAI/2.0/oai_dc/" xsi:schemaLocation="http://www.openarchives.org/OAI/2.0/oai_dc/ http://www.openarchives.org/OAI/2.0/oai_dc.xsd">
-<dc:title>Project 0-1787 Seal Coat Constructability Review</dc:title>
-<dc:creator>Gransburg, Douglas</dc:creator>
-<dc:creator>Senadheera, Sanjaya</dc:creator>
-<dc:description>This research project conducted a
-formal constructability review of
-TxDOT seal coat practices. The
-essence of any “constructability
-review” is the identification of best
-practices. It is a technique that is
-being widely implemented by the
-federal government, state DOTs and
-other public agencies.
-The researchers picked apart, piece
-by piece, the seal coat process from
-planning to construction
-completion, looking for those
-portions of the process that are
-inherently variable and difficult to
-replicate in the field. The
-performance of a seal coat project is
-influenced by a number of factors.</dc:description>
-<dc:date>2010-11-30T20:09:05Z</dc:date>
-<dc:date>2012-05-03T19:11:49Z</dc:date>
-<dc:date>2010-11-30T20:09:05Z</dc:date>
-<dc:date>2012-05-03T19:11:49Z</dc:date>
-<dc:date>2004</dc:date>
-<dc:type>Technical Report</dc:type>
-<dc:identifier>http://hdl.handle.net/2346/2100</dc:identifier>
-<dc:language>en_US</dc:language>
-<dc:relation>;0-1787</dc:relation>
+Example Record
+
+<oai_dc:dc xmlns:dc="http://purl.org/dc/elements/1.1/"
+    xmlns:doc="http://www.lyncode.com/xoai"
+    xmlns:oai_dc="http://www.openarchives.org/OAI/2.0/oai_dc/"
+    xsi:schemaLocation="http://www.openarchives.org/OAI/2.0/oai_dc/
+    http://www.openarchives.org/OAI/2.0/oai_dc.xsd">
+    <dc:title>Project 0-1787 Seal Coat Constructability Review</dc:title>
+    <dc:creator>Gransburg, Douglas</dc:creator>
+    <dc:creator>Senadheera, Sanjaya</dc:creator>
+    <dc:description>This research project conducted a
+        formal constructability review of
+        TxDOT seal coat practices. The
+        essence of any “constructability
+        review” is the identification of best
+        practices. It is a technique that is
+        being widely implemented by the
+        federal government, state DOTs and
+        other public agencies.
+        The researchers picked apart, piece
+        by piece, the seal coat process from
+        planning to construction
+        completion, looking for those
+        portions of the process that are
+        inherently variable and difficult to
+        replicate in the field. The
+        performance of a seal coat project is
+        influenced by a number of factors.
+    </dc:description>
+    <dc:date>2010-11-30T20:09:05Z</dc:date>
+    <dc:date>2012-05-03T19:11:49Z</dc:date>
+    <dc:date>2010-11-30T20:09:05Z</dc:date>
+    <dc:date>2012-05-03T19:11:49Z</dc:date>
+    <dc:date>2004</dc:date>
+    <dc:type>Technical Report</dc:type>
+    <dc:identifier>http://hdl.handle.net/2346/2100</dc:identifier>
+    <dc:language>en_US</dc:language>
+    <dc:relation>;0-1787</dc:relation>
 </oai_dc:dc>
 """
