+++ providers/pe/upc/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'providers.pe.upc.apps.AppConfig'
